AC: move convenience functions to instructions
