+++ account-compression/sdk/src/index.ts
@@ -3,4 +3,3 @@ export * from './instructions';
 export * from './accounts';
 export { PathNode } from './accounts';
 export * from './utils';
-export * from './convenience';
